refactor clear facet, https://github.com/TheScienceMuseum/collectionsonline/pull/223#discussion_r72763298
