+++ lib/transforms/search-results-to-template-data.js
@@ -5,7 +5,7 @@ const getValues = require('./get-values');
 const getPrimaryValue = require('../get-primary-value');
 const createSelectedFilters = require('./create-selected-filters');
 const getNestedProperty = require('../nested-property');
-const queryParamsPerCategory = require('../query-params/query-params-per-category');
+const createClearFacetLinks = require('./create-clear-facet-links');
 
 module.exports = (queryParams, results, config) => {
   config = config || {};
@@ -26,7 +26,7 @@ module.exports = (queryParams, results, config) => {
       filterPeople: createTypeFilterLink('people', queryParams, rootUrl),
       filterDocuments: createTypeFilterLink('documents', queryParams, rootUrl),
       clearFilter: createClearFilterLink(queryParams, rootUrl),
-      clearFacet: createClearFacetLink(queryParams, rootUrl),
+      clearFacet: createClearFacetLinks(queryParams, rootUrl),
       grid: createTypePageLink('search', queryParams, rootUrl),
       list: createTypePageLink('results-list', queryParams, rootUrl)
     }
@@ -140,34 +140,4 @@ function createClearFilterLink (queryParams, rootUrl) {
   };
   params['page[type]'] = queryParams.pageType;
   return `${rootUrl}/search${type}?${QueryString.stringify(params)}`;
-}
-
-function createClearFacetLink (queryParams, rootUrl) {
-  // queryParams.query contain already the query {"q":"ada","page[size]":50,"page[type]":"search"}
-  const links = {};
-  const type = queryParams.type !== 'all' ? '/' + queryParams.type : '';
-  const facetsName = queryParamsPerCategory[queryParams.type];
-  facetsName.forEach(function (facet) {
-    links[facet] = `${rootUrl}/search${type}?${QueryString.stringify(ignoreQueryParam(facet, queryParams.query))}`;
-  });
-  // special links for the date: need to ignore the both date at the same time
-  var queryPramsNoDates = ignoreQueryParam('filter[date[from]]', queryParams.query);
-  queryPramsNoDates = ignoreQueryParam('filter[date[to]]', queryPramsNoDates);
-  queryPramsNoDates = ignoreQueryParam('filter[birth[date]]', queryPramsNoDates);
-  queryPramsNoDates = ignoreQueryParam('filter[death[date]]', queryPramsNoDates);
-  links['dates'] = `${rootUrl}/search${type}?${QueryString.stringify(queryPramsNoDates)}`;
-  return links;
-}
-
-function ignoreQueryParam (param, queryParams) {
-  const result = {};
-  const keys = Object.keys(queryParams).filter(function (k) {
-    return k !== param;
-  });
-
-  keys.forEach(function (k) {
-    result[k] = queryParams[k];
-  });
-
-  return result;
 }
