+++ package.json
@@ -62,7 +62,7 @@
     "test-all": "run-s test:*",
     "test:lint": "semistandard",
     "test:unit": "istanbul cover tape 'test/**/*.test.js'",
-    "test:endtoend": "nightwatch --retries 3 --env ${NW_ENV}",
+    "test-endtoend": "nightwatch --retries 3 --env ${NW_ENV}",
     "build": "run-p build:*",
     "build:js": "browserify client/main.js -o public/bundle.js",
     "build:css": "run-s build:css:*",
